+++ src/10_styles/02_molecule-styles/_m-ex-link.scss
@@ -1,4 +1,4 @@
-.a-ex-link {
+.m-ex-link {
     color: $link-color;
 
     svg {
