+++ src/10_styles/styles.scss
@@ -10,9 +10,9 @@
 // Atom Styles
 @import "01_atom-styles/a-icon";
 @import "01_atom-styles/skip-link";
-@import "01_atom-styles/a-ex-link";
 
 // Molecule Styles
+@import "02_molecule-styles/m-ex-link";
 @import "02_molecule-styles/m-nav-links";
 
 // Organism Styles
